+++ firmware/src/modes/BreakoutClockMode.cpp
@@ -48,7 +48,6 @@ void BreakoutClockMode::handle()
 #endif // EXTENSION_MICROPHONE
     const uint8_t nextX{
         static_cast<uint8_t>(lroundf(xDec + (cosf(static_cast<float>(deg) * static_cast<float>(DEG_TO_RAD)) * speed)))};
-    // NOLINTNEXTLINE(cppcoreguidelines-init-variables)
     const uint8_t nextY{static_cast<uint8_t>(
         std::lroundf(yDec - (sinf(static_cast<float>(deg) * static_cast<float>(DEG_TO_RAD)) * speed)))};
     if (y == 0U && deg < 180U)
@@ -64,7 +63,7 @@ void BreakoutClockMode::handle()
     if (nextY >= GRID_ROWS - 1U && deg > 180U)
     {
         // Bottom
-        deg = random(30, 151); // ±60°
+        deg = static_cast<uint16_t>(random(30, 151)); // ±60°
     }
     else if ((nextX != x || nextY != y) && Display.getPixel(nextX, nextY) != 0U)
     {
